+++ node/commons/compatibility-layer.ts
@@ -1,4 +1,5 @@
 import { distinct } from '../utils/object'
+import unescape from 'unescape'
 
 export enum IndexingType {
   API = 'API',
@@ -35,6 +36,13 @@ export const convertBiggyProduct = (
 
   const brandId = product.brandId ? Number(product.brandId) : -1
 
+  const selectedProperties = product.split && [
+    {
+      key: product.split.labelKey,
+      value: product.split.labelValue,
+    },
+  ]
+
   const convertedProduct: SearchProduct & { cacheId?: string, [key: string]: any } = {
     categories,
     categoriesIds,
@@ -62,6 +70,7 @@ export const convertBiggyProduct = (
     itemMetadata: {
       items: []
     },
+    selectedProperties,
   }
 
   if (product.extraData) {
@@ -298,7 +307,8 @@ export const convertOrderBy = (orderBy?: string): string => {
 
 export const buildBreadcrumb = (
   attributes: ElasticAttribute[],
-  fullText: string
+  fullText: string,
+  selectedFacets: SelectedFacet[]
 ) => {
   const pivotValue: string[] = []
   const pivotMap: string[] = []
@@ -316,20 +326,43 @@ export const buildBreadcrumb = (
   }
 
   const activeAttributes = attributes.filter(attribute => attribute.active)
+  const activeValues: (ElasticAttributeValue & {
+    visible: boolean
+    attributeKey: string
+  })[] = []
 
-  activeAttributes.map(attribute => {
+  activeAttributes.forEach(attribute => {
     attribute.values.forEach(value => {
-      if (!value.active) {
-        return
+      if (value.active) {
+        activeValues.push({
+          ...value,
+          visible: attribute.visible,
+          attributeKey: attribute.key,
+        })
       }
+    })
+  })
 
-      pivotValue.push(value.key)
-      pivotMap.push(attribute.key)
+  const selectedFacetsValues = selectedFacets.map(
+    selectedFacet => selectedFacet.value
+  )
+  activeValues.sort((a, b) =>
+    selectedFacetsValues.indexOf(a.key) < selectedFacetsValues.indexOf(b.key)
+      ? -1
+      : 1
+  )
 
-      breadcrumb.push({
-        name: unescape(value.label),
-        href: `/${pivotValue.join('/')}?map=${pivotMap.join(',')}`,
-      })
+  activeValues.forEach(value => {
+    pivotValue.push(value.key)
+    pivotMap.push(value.attributeKey)
+
+    if (!value.visible) {
+      return
+    }
+
+    breadcrumb.push({
+      name: unescape(value.label),
+      href: `/${pivotValue.join('/')}?map=${pivotMap.join(',')}`,
     })
   })